Move output shape generation and error generation to AWSResponse
